Feature/navigation UI (#35)
* Addition of a getTabTestTag function"

Addition of the new function getTabTestTag used in BottomNavigationMenu.kt for testing purposes

* Creation of the new kotlin file BottomNavigationMenu.kt, used for the UI of the bottom navigation bar.

This new file is used for the UI of the navigation bar based on the figma, all the UI is as showed in the figma except the selection box color to be changed.

* Tests for the UI of the navigation bar

Some tests that tests if the UI for the navigation bar works on itself (clicks, view, etc.) no viewModel tests.

* fix(navigation): reformated code

* refactor(navigation): update NavigationTestUI to use BottomNavigationMenu

* refactor(navigation): format code in NavigationTestUI

---------

Co-authored-by: chibo <erwan.guesty@gmail.com>
Co-authored-by: Maxime-T <maxime.tscharner@gmail.com>