Feature/top app bar (#74)
* Addition of a getTabTestTag function"

Addition of the new function getTabTestTag used in BottomNavigationMenu.kt for testing purposes

* Creation of the new kotlin file BottomNavigationMenu.kt, used for the UI of the bottom navigation bar.

This new file is used for the UI of the navigation bar based on the figma, all the UI is as showed in the figma except the selection box color to be changed.

* Tests for the UI of the navigation bar

Some tests that tests if the UI for the navigation bar works on itself (clicks, view, etc.) no viewModel tests.

* fix(navigation): reformated code

* refactor(navigation): update NavigationTestUI to use BottomNavigationMenu

* refactor(navigation): format code in NavigationTestUI

* UI for the topAppBar

UI for the top App Bar showing the title of the screen and a button to access the profile.

* test for UI topAppBar

Exhaustive tests for the UI topAppBar to see if basic actions can be performed

* added test type

added 3 tests type for the new UI topBarApp tests
TopNavigationBar, SettingsButton, MenuButton

* fix(topAppBar): fixed formatting

* Delete app/src/androidTest/java/com/android/sample/navigation/NavigationTestUI.kt

Removed a file that was already in main, but was also in this PR.

* Delete app/src/main/java/com/android/sample/ui/navigation/BottomNavigationMenu.kt

Removed this file because it was already in main and he was also in this PR

* feat(navigation): implement TopNavigationBar and update navigation logic

---------

Co-authored-by: chibo <erwan.guesty@gmail.com>
Co-authored-by: Maxime-T <maxime.tscharner@gmail.com>
Co-authored-by: Lucas Guesry <144180129+LucasGuesry@users.noreply.github.com>